--- CHANGELOG.md
@@ -1,17 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
-
-### [0.0.3](https://github.com/minjs1cn/vite-plugin-best-mock/compare/v0.0.2...v0.0.3) (2022-10-20)
-
-### 0.0.2 (2022-10-20)
-
-
-### Features
-
-* 开发环境支持完成 ([4202b99](https://github.com/minjs1cn/vite-plugin-best-mock/commit/4202b9982222224cb57383a77356a7b8739efdd3))
-* 抢占名称 ([2278c72](https://github.com/minjs1cn/vite-plugin-best-mock/commit/2278c72ebcf21f4587ff9e7b8ebc15c472de9f5a))
-* 抢占名称 ([44e189e](https://github.com/minjs1cn/vite-plugin-best-mock/commit/44e189ea9cc0a6e41758c6120bc5de3d9713c486))
-* iniit ([21c8e6e](https://github.com/minjs1cn/vite-plugin-best-mock/commit/21c8e6e70cfaed065ea89e7cd22a64a53a688c2e))
-* init ([74b9d9c](https://github.com/minjs1cn/vite-plugin-best-mock/commit/74b9d9c70650366d5b53f95c196cb1d0ecfc4b71))
-* logger ([f525c2c](https://github.com/minjs1cn/vite-plugin-best-mock/commit/f525c2cc1193be3b1407c3fd98d088627d0483e2))
