+++ package.json
@@ -1,59 +1,16 @@
 {
   "name": "vite-plugin-best-mock",
-  "version": "0.0.5",
-  "description": "A mock Plugin for vite",
-  "main": "./lib/index.js",
+  "private": true,
+  "repository": "git@github.com:minjs1cn/vite-plugin-best-mock.git",
   "license": "MIT",
-  "author": "qingyang",
-  "files": [
-    "lib"
-  ],
-  "type": "module",
-  "engines": {
-    "node": "^14.20.0"
-  },
-  "keywords": [
-    "vite",
-    "plugin",
-    "typescript",
-    "mock",
-    "middleware",
-    "connect",
-    "multiparty"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/minjs1cn/vite-plugin-best-mock"
-  },
-  "bugs": {
-    "url": "https://github.com/minjs1cn/vite-plugin-best-mock/issues",
-    "email": "303890562@qq.com"
-  },
-  "homepage": "https://github.com/minjs1cn/vite-plugin-best-mock/tree/main/#readme",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc",
-    "preview": "vite build",
-    "release": "standard-version"
-  },
-  "peerDependencies": {
-    "mockjs": ">=1.1.0",
-    "vite": ">=2.0.0"
+    "dev": "pnpm run dev --filter vite-plugin-best-mock",
+    "build": "pnpm run build --filter vite-plugin-best-mock",
+    "preview": "pnpm run build --filter playground",
+    "playground": "pnpm run dev --filter playground",
+    "pre": "pnpm add vite-plugin-best-mock --filter playground"
   },
   "dependencies": {
-    "chalk": "^5.1.2",
-    "colors": "^1.4.0",
-    "connect-multiparty": "^2.2.0",
-    "qs": "^6.11.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^4.8.4",
-    "vite": "^3.1.8"
-  },
-  "devDependencies": {
-    "@types/multiparty": "^0.0.33",
-    "@types/node": "^18.11.0",
-    "@types/qs": "^6.9.7",
-    "pkg-get": "^0.0.2",
-    "standard-version": "^9.5.0"
+    "tsup": "^6.3.0"
   }
 }
