+++ packages/vite-plugin-best-mock/package.json
@@ -1,0 +1,65 @@
+{
+  "name": "vite-plugin-best-mock",
+  "version": "0.0.5",
+  "description": "A mock Plugin for vite",
+  "main": "./dist/index.cjs",
+  "type": "module",
+  "license": "MIT",
+  "author": "qingyang",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
+    }
+  },
+  "types": "./dist/index.d.ts",
+  "files": [
+    "lib"
+  ],
+  "engines": {
+    "node": "^14.20.0"
+  },
+  "keywords": [
+    "vite",
+    "plugin",
+    "typescript",
+    "mock",
+    "middleware",
+    "connect",
+    "multiparty"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/minjs1cn/vite-plugin-best-mock"
+  },
+  "bugs": {
+    "url": "https://github.com/minjs1cn/vite-plugin-best-mock/issues",
+    "email": "303890562@qq.com"
+  },
+  "homepage": "https://github.com/minjs1cn/vite-plugin-best-mock/tree/main/#readme",
+  "scripts": {
+    "dev": "tsup --watch",
+    "build": "tsup",
+    "release": "standard-version"
+  },
+  "peerDependencies": {
+    "vite": ">=2.0.0"
+  },
+  "dependencies": {
+    "chalk": "^5.1.2",
+    "colors": "^1.4.0",
+    "connect-multiparty": "^2.2.0",
+    "qs": "^6.11.0",
+    "ts-node": "^10.9.1"
+  },
+  "devDependencies": {
+    "@types/multiparty": "^0.0.33",
+    "@types/node": "^18.11.0",
+    "@types/qs": "^6.9.7",
+    "pkg-get": "^0.0.2",
+    "standard-version": "^9.5.0",
+    "typescript": "^4.8.4",
+    "vite": "^3.1.8",
+    "tsup": "^6.3.0"
+  }
+}
