+++ packages/vite-plugin-best-mock/src/connect-multiparty.d.ts
