+++ packages/vite-plugin-best-mock/tsup.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  entry: ["src/index.ts"],
+  dts: true,
+  outDir: "dist",
+  format: ["cjs", "esm"],
+  clean: true,
+  sourcemap: true,
+});
