+++ playground/mock/user/[id].ts
