+++ playground/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "playground",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "typescript": "^4.6.4",
+    "vite": "^3.2.0",
+    "vite-plugin-best-mock": "workspace:^0.0.5"
+  }
+}
